use settins , profile web route and controller
